Half_don

+++ enemies.py
@@ -1,4 +1,7 @@
 
+from sys import float_repr_style
+
+
 class Enemies:
     Name: str
     hp: int
@@ -25,11 +28,11 @@ class Enemymods:
     file_name: str
     file_hp: float
     file_atk: float
-    file_spd: int
+    file_spd: float
 
     def __init__(self, fsor: str):
         file = fsor.split(' ')
         self.file_name = file[0]
         self.file_hp = float(file[1])
         self.file_atk = float(file[2])
-        self.file_spd = int(file[3])
+        self.file_spd = float(file[3])
